requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.4
django-ckeditor==6.7.3
django-crispy-forms==2.4
django-js-asset==3.1.2
sqlparse==0.5.3
tzdata==2025.2
